Add SimpleQueue class and SimpleQueueTest
Also switched to eclipse Luna, since Ant 1.8 was having problems with
java 8, and it didn't look like there was a way to upgrade the eclipse
version of ant for Kepler. I'll add some sort of README for how to set
up the dev environment to be similar to what I have.