Add 3 unit tests for sorting Point2Ds with PolarSort
Test 1 checks that if all points are non-zero and different direction,
that sorting works, on all possible permutations on array sizes from 1
to 8.

Test 2 checks that if all points are in the same direction, that sorting
works.  This includes a point at the origin.

Test 3 checks that if we sort 1,000,000 points, it doesn't take too
long.